+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Configurable key bindings
 - Horizontal scrolling
 
+### Fixed
+- A noop rebase will no longer return a non-zero status code
+
 ## [1.0.0] - 2019-04-10
 
 ### Added